// app/services/olxApiClient/api/schema.ts
export const CategoryTypeSchema = z.string()
export type CategoryType = z.infer<typeof CategoryTypeSchema>

export const ModeSchema = z.string()
export type Mode = z.infer<typeof ModeSchema>
// "builttype",
//     "floor_select",
//     "furniture",
//     "m",
//     "market",
//     "price",
//     "price_per_m",
//     "rent",
//     "rooms",
export const KeySchema = z.enum([
  'builttype',
  'floor_select',
  'furniture',
  'm',
  'price',
  'price_per_m',
  'market',
  'rent',
  'rooms',
  'roomsize',
  'preferences',
])
export type Key = z.infer<typeof KeySchema>

export const CityNameSchema = z.string()
export type CityName = z.infer<typeof CityNameSchema>

export const NormalizedNameSchema = z.string()
export type NormalizedName = z.infer<typeof NormalizedNameSchema>

export const OfferTypeSchema = z.enum(['offer'])
// ...
export const DatumStatusSchema = z.enum([
  'active',
  'removed_by_user',
  'disabled',
  'removed_by_moderator',
])
export type DatumStatus = z.infer<typeof DatumStatusSchema>

export const CompanyNameSchema = z.string()
// ...
export const UserSchema = z.object({
  id: z.number(),
  created: z.string(),
  other_ads_enabled: z.boolean(),
  name: z.string(),
  logo: z.union([z.null(), z.string()]),
  photo: z.union([z.null(), z.string()]),
  banner_mobile: z.string(),
  banner_desktop: z.string(),
  company_name: CompanyNameSchema,
  about: z.string(),
  b2c_business_page: z.boolean(),
  is_online: z.boolean(),
  last_seen: z.string(),
  seller_type: z.null(),
  uuid: z.string(),
})
// ...
export const PhotoSchema = z.object({
  id: z.number(),
  filename: z.string(),
  rotation: z.number(),
  width: z.number(),
  height: z.number(),
  link: z.string(),
})
// ...
const AreaParam = z.object({
  key: z.literal('m'),
  name: z.string(),
  type: z.literal('input'),
  value: z.object({
    key: z.union([z.null(), z.string()]).optional(),
    value: z.union([z.number(), z.null()]).optional(),
  }),
})

const RentParam = z.object({
  key: z.literal('rent'),
  name: z.string(),
  type: z.literal('input'),
  value: z.object({
    key: z.union([z.null(), z.string()]).optional(),
    label: z.string(),
  }),
})

const PriceParam = z.object({
  key: z.literal('price'),
  name: z.string(),
  type: z.literal('price'),
  value: z.object({
    value: z.union([z.number(), z.null()]).optional(),
    currency: z.union([CurrencySchema, z.null()]).optional(),
    negotiable: z.union([z.boolean(), z.null()]).optional(),
    label: z.string(),
  }),
})

const FurnitureParam = z.object({
  key: z.literal('furniture'),
  name: z.string(),
  type: z.literal('select'),
  value: z.object({
    key: z.union([z.null(), z.string()]).optional(),
    label: z.string(),
  }),
})

const RoomsParam = z.object({
  key: z.literal('rooms'),
  name: z.string(),
  type: z.literal('select'),
  value: z.object({
    key: z.union([z.null(), z.string()]).optional(),
    label: z.string(),
  }),
})

const BuiltTypeParam = z.object({
  key: z.literal('builttype'),
  name: z.string(),
  type: z.literal('select'),
  value: z.object({
    key: z.union([z.null(), z.string()]).optional(),
    label: z.string(),
  }),
})

const FloorParam = z.object({
  key: z.literal('floor_select'),
  name: z.string(),
  type: z.literal('select'),
  value: z.object({
    key: z.union([z.null(), z.string()]).optional(),
    label: z.string(),
  }),
})

const RoomSizeParam = z.object({
  key: z.literal('roomsize'),
  name: z.string(),
  type: z.literal('select'),
  value: z.object({
    key: z.union([z.null(), z.string()]).optional(),
    label: z.string(),
  }),
})

const PreferencesParam = z.object({
  key: z.literal('preferences'),
  name: z.string(),
  type: z.literal('select'),
  value: z.object({
    key: z.union([z.null(), z.string()]).optional(),
    label: z.string(),
  }),
})

const ParkingParam = z.object({
  key: z.literal('parking'),
  name: z.string(),
  type: z.literal('checkboxes'),
  value: z.object({
    key: z.array(z.string()),
    label: z.string(),
  }),
})

const ElevatorParam = z.object({
  key: z.literal('winda'),
  name: z.string(),
  type: z.literal('select'),
  value: z.object({
    key: z.union([z.null(), z.string()]).optional(),
    label: z.string(),
  }),
})

const PricePerM = z.object({
  key: z.literal('price_per_m'),
  name: z.string(),
  type: z.literal('input'),
  value: z.object({
    value: z.union([z.number(), z.null()]).optional(),
    currency: z.union([CurrencySchema, z.null()]).optional(),
    negotiable: z.union([z.boolean(), z.null()]).optional(),
    label: z.string(),
  }),
})

const Market = z.object({
  key: z.literal('market'),
  name: z.string(),
  type: z.literal('select'),
  value: z.object({
    key: z.union([z.null(), z.string()]).optional(),
    label: z.string(),
  }),
})

const PetsParam = z.object({
  key: z.literal('pets'),
  name: z.string(),
  type: z.literal('select'),
  value: z.object({
    key: z.union([z.null(), z.string()]).optional(),
    label: z.string(),
  }),
})

export const ParamSchema = z.discriminatedUnion('key', [
  AreaParam,
  RentParam,
  PriceParam,
  FurnitureParam,
  RoomsParam,
  BuiltTypeParam,
  FloorParam,
  RoomSizeParam,
  PreferencesParam,
  PricePerM,
  Market,
  ParkingParam,
  ElevatorParam,
  PetsParam,
])
export type Param = z.infer<typeof ParamSchema>

export const MapSchema = z.object({
  zoom: z.number(),
  lat: z.number(),
  lon: z.number(),
  radius: z.number(),
  show_detailed: z.boolean(),
})
export type Map = z.infer<typeof MapSchema>

export const DistrictSchema = z.object({
  id: z.number(),
  name: z.string(),
})
export type District = z.infer<typeof DistrictSchema>

export const CitySchema = z.object({
  id: z.number(),
  name: CityNameSchema,
  normalized_name: NormalizedNameSchema,
})
export type City = z.infer<typeof CitySchema>

export const LocationSchema = z.object({
  city: CitySchema,
  district: z.union([DistrictSchema, z.null()]).optional(),
  region: CitySchema,
})
export type Location = z.infer<typeof LocationSchema>

export const RockSchema = z.object({
  offer_id: z.null(),
  active: z.boolean(),
  mode: ModeSchema,
})
export type Rock = z.infer<typeof RockSchema>

export const DeliverySchema = z.object({
  rock: RockSchema,
})
export type Delivery = z.infer<typeof DeliverySchema>

export const ContactSchema = z.object({
  name: z.string(),
  phone: z.boolean(),
  chat: z.boolean(),
  negotiation: z.boolean(),
  courier: z.boolean(),
})
export type Contact = z.infer<typeof ContactSchema>

export const CategorySchema = z.object({
  id: z.number(),
  type: CategoryTypeSchema,
})
export type Category = z.infer<typeof CategorySchema>

export const OlxOfferSchema = z.object({
  id: z.number(),
  url: z.string(),
  title: z.string(),
  last_refresh_time: z.string(),
  created_time: z.string(),
  valid_to_time: z.string(),
  pushup_time: z.union([z.null(), z.string()]),
  description: z.string(),
  params: z.array(ParamSchema),
  key_params: z.array(KeySchema),
  business: z.boolean(),
  user: UserSchema,
  status: DatumStatusSchema,
  contact: ContactSchema,
  map: MapSchema,
  location: LocationSchema,
  photos: z.array(PhotoSchema),
  category: CategorySchema,
  offer_type: OfferTypeSchema,
  external_url: z.union([z.null(), z.string()]).optional(),
})
export type Datum = z.infer<typeof OlxOfferSchema>

export const OlxApiResponseSchema = z.object({
  data: z.array(z.record(z.unknown())),
})